requirements.txt
streamlit
streamlit-chat
PyPDF2
sentence-transformers
chromadb
Pillow
ollama
